+++ fdbserver/GrvProxyServer.actor.cpp
@@ -920,12 +920,13 @@ ACTOR static Future<Void> transactionStarter(GrvProxyInterface proxy,
 
 ACTOR Future<Void> grvProxyServerCore(GrvProxyInterface proxy,
                                       MasterInterface master,
+									  LifetimeToken masterLifetime,
                                       Reference<AsyncVar<ServerDBInfo> const> db) {
 	state GrvProxyData grvProxyData(proxy.id(), master, proxy.getConsistentReadVersion, db);
 
 	state PromiseStream<Future<Void>> addActor;
 	state Future<Void> onError =
-	    transformError(actorCollection(addActor.getFuture()), broken_promise(), master_tlog_failed());
+	    transformError(actorCollection(addActor.getFuture()), broken_promise(), tlog_failed());
 
 	state GetHealthMetricsReply healthMetricsReply;
 	state GetHealthMetricsReply detailedHealthMetricsReply;
@@ -933,9 +934,14 @@ ACTOR Future<Void> grvProxyServerCore(GrvProxyInterface proxy,
 	addActor.send(waitFailureServer(proxy.waitFailure.getFuture()));
 	addActor.send(traceRole(Role::GRV_PROXY, proxy.id()));
 
+	TraceEvent("GrvProxyServerCore", proxy.id())
+	    .detail("MasterId", master.id().toString())
+	    .detail("MasterLifetime", masterLifetime.toString())
+	    .detail("RecoveryCount", db->get().recoveryCount);
+
 	// Wait until we can load the "real" logsystem, since we don't support switching them currently
-	while (!(grvProxyData.db->get().master.id() == master.id() &&
-	         grvProxyData.db->get().recoveryState >= RecoveryState::RECOVERY_TRANSACTION)) {
+	while (!(masterLifetime.isEqual(grvProxyData.db->get().masterLifetime) &&
+	         grvProxyData.db->get().recoveryState >= RecoveryState::ACCEPTING_COMMITS)) {
 		wait(grvProxyData.db->onChange());
 	}
 	// Do we need to wait for any db info change? Yes. To update latency band.
@@ -956,7 +962,7 @@ ACTOR Future<Void> grvProxyServerCore(GrvProxyInterface proxy,
 		when(wait(dbInfoChange)) {
 			dbInfoChange = grvProxyData.db->onChange();
 
-			if (grvProxyData.db->get().master.id() == master.id() &&
+			if (masterLifetime.isEqual(grvProxyData.db->get().masterLifetime) &&
 			    grvProxyData.db->get().recoveryState >= RecoveryState::RECOVERY_TRANSACTION) {
 				grvProxyData.logSystem =
 				    ILogSystem::fromServerDBInfo(proxy.id(), grvProxyData.db->get(), false, addActor);
@@ -983,13 +989,13 @@ ACTOR Future<Void> grvProxyServer(GrvProxyInterface proxy,
                                   InitializeGrvProxyRequest req,
                                   Reference<AsyncVar<ServerDBInfo> const> db) {
 	try {
-		state Future<Void> core = grvProxyServerCore(proxy, req.master, db);
+		state Future<Void> core = grvProxyServerCore(proxy, req.master, req.masterLifetime, db);
 		wait(core || checkRemoved(db, req.recoveryCount, proxy));
 	} catch (Error& e) {
 		TraceEvent("GrvProxyTerminated", proxy.id()).error(e, true);
 
 		if (e.code() != error_code_worker_removed && e.code() != error_code_tlog_stopped &&
-		    e.code() != error_code_master_tlog_failed && e.code() != error_code_coordinators_changed &&
+		    e.code() != error_code_tlog_failed && e.code() != error_code_coordinators_changed &&
 		    e.code() != error_code_coordinated_state_conflict && e.code() != error_code_new_coordinators_timed_out) {
 			throw;
 		}
