+++ flow/Trace.h
@@ -599,5 +599,8 @@ extern TraceBatch g_traceBatch;
 #define DUMPTOKEN(name)                                                                                                \
 	TraceEvent("DumpToken", recruited.id()).detail("Name", #name).detail("Token", name.getEndpoint().token)
 
+#define DUMPTOKEN_PROCESS(process, name)                                                                               \
+	TraceEvent("DumpToken", process.id()).detail("Name", #name).detail("Token", name.getEndpoint().token)	
+
 #define DisabledTraceEvent(...) false && TraceEvent()
 #endif
